requirements.txt
dotenv
AzureOpenAI